Time: 3 ms (67.43%) | Memory: 101.4 MB (78.45%) - LeetSync

+++ 2358-number-of-ways-to-split-array/number-of-ways-to-split-array.java
@@ -9,7 +9,7 @@ class Solution {
         }
         int c=0;
         for(int j=0;j<n-1;j++){
-            if((long)prefix[j]>=(long)(prefix[n-1]-prefix[j])) c++;
+            if(prefix[j]>=(prefix[n-1]-prefix[j])) c++;
         }
     return c;
     }
